refactor: extract room dashboard component and consolidate home actions
